assignment/user_authentication_system/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.7
alembic==1.15.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.9.1
python-multipart==0.0.20
jinja2==3.1.6
email-validator==2.2.0
pytest==8.3.5
httpx==0.28.1
bcrypt==4.0.1
